Move updatable actionable types to /models
