+++ src/models/actionable/index.ts
@@ -11,6 +11,9 @@ export * from './creatable-actionable';
 export * from './is-actionable';
 export * from './raw-actionable';
 export * from './raw-creatable-actionable';
+export * from './raw-updatable-actionable';
 export * from './to-actionable';
 export * from './to-raw-actionable';
 export * from './to-raw-creatable-actionable';
+export * from './to-raw-updatable-actionable';
+export * from './updatable-actionable';
