+++ src/models/actionable/raw-updatable-actionable.ts
@@ -1,0 +1,21 @@
+import { RawNumericID, RawUUID } from '../../value-objects';
+import { RawActionableAction, RawActionableTrigger } from './raw-actionable';
+
+export interface RawUpdatableActionable {
+	// ?Question: why do we allow that field to be set?
+	GUID?: RawUUID;
+
+	UID: RawNumericID;
+	GIDs: Array<RawNumericID>;
+
+	Global: boolean;
+	Labels: Array<string>;
+
+	Name: string;
+	Description: string | null; // Null becomes empty string
+	Contents: {
+		menuLabel: string | null;
+		actions: Array<RawActionableAction>;
+		triggers: Array<RawActionableTrigger>;
+	};
+}
